requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.7
djangorestframework==3.15.2
filelock==3.15.4
mysqlclient==2.2.4
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.3
